#6 Updated build tools to wordpress/scripts 5.0.1
