--- .babelrc
@@ -1,7 +1,0 @@
-{
-	"plugins": [
-		["@babel/plugin-transform-react-jsx", {
-			"pragma": "wp.element.createElement"
-		}]
-	]
-}
