+++ package.json
@@ -2,18 +2,14 @@
   "name": "accordion-blocks",
   "version": "1.0.2",
   "description": "Gutenberg blocks for creating responsive accordion drop-downs.",
-  "main": "js/accordion-blocks.js",
+  "main": "src/accordion-blocks.js",
   "scripts": {
-    "dev": "webpack --watch",
-    "build": "NODE_ENV=production webpack"
+    "start": "wp-scripts start",
+    "build": "wp-scripts build"
   },
   "author": "pbuchanan",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "7.3.4",
-    "@babel/plugin-transform-react-jsx": "7.3.0",
-    "babel-loader": "8.0.5",
-    "webpack": "4.29.6",
-    "webpack-cli": "3.2.3"
+    "@wordpress/scripts": "5.0.1"
   }
 }
